requirements.txt
numpy==1.24.4
torch==2.0.1+cpu
torchaudio==2.0.2+cpu
transformers==4.33.2
tokenizers==0.13.3
sentence-transformers==2.2.2
streamlit==1.28.1
faiss-cpu==1.7.4
pandas==2.0.3
openai==1.3.7
python-dotenv==1.0.0
huggingface-hub==0.16.4
accelerate==0.21.0
scikit-learn==1.3.2
